Extract common CommandMarker code to helper class
