+++ driver-core/src/main/com/mongodb/internal/async/client/CommandMarker.java
@@ -16,22 +16,18 @@
 
 package com.mongodb.internal.async.client;
 
-import com.mongodb.Block;
-import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientException;
-import com.mongodb.MongoClientSettings;
 import com.mongodb.MongoTimeoutException;
 import com.mongodb.ReadConcern;
 import com.mongodb.ReadPreference;
-import com.mongodb.connection.ClusterSettings;
 import com.mongodb.internal.async.SingleResultCallback;
 import org.bson.RawBsonDocument;
 
 import java.io.Closeable;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 
-import static com.mongodb.internal.capi.MongoCryptOptionsHelper.createMongocryptdSpawnArgs;
+import static com.mongodb.internal.capi.MongoCryptHelper.createMongocryptdSpawnArgs;
+import static com.mongodb.internal.capi.MongoCryptHelper.createMongocryptdClientSettings;
 
 @SuppressWarnings("UseOfProcessBuilder")
 class CommandMarker implements Closeable {
@@ -50,24 +46,7 @@ class CommandMarker implements Closeable {
             } else {
                 processBuilder = null;
             }
-
-            String connectionString;
-            if (options.containsKey("mongocryptdURI")) {
-                connectionString = (String) options.get("mongocryptdURI");
-            } else {
-                connectionString = "mongodb://localhost:27020";
-            }
-
-            client = AsyncMongoClients.create(MongoClientSettings.builder()
-                .applyConnectionString(new ConnectionString(connectionString))
-                .applyToClusterSettings(new Block<ClusterSettings.Builder>() {
-                    @Override
-                    public void apply(final ClusterSettings.Builder builder) {
-                        builder.serverSelectionTimeout(1, TimeUnit.SECONDS);
-                    }
-                })
-                .build());
-
+            client = AsyncMongoClients.create(createMongocryptdClientSettings((String) options.get("mongocryptdURI")));
         }
     }
