+++ driver-core/src/main/com/mongodb/internal/async/client/Crypts.java
@@ -26,7 +26,7 @@ import com.mongodb.crypt.capi.MongoCrypts;
 import javax.net.ssl.SSLContext;
 import java.security.NoSuchAlgorithmException;
 
-import static com.mongodb.internal.capi.MongoCryptOptionsHelper.createMongoCryptOptions;
+import static com.mongodb.internal.capi.MongoCryptHelper.createMongoCryptOptions;
 
 public final class Crypts {
