+++ driver-sync/src/main/com/mongodb/client/internal/CommandMarker.java
@@ -16,25 +16,21 @@
 
 package com.mongodb.client.internal;
 
-import com.mongodb.Block;
-import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientException;
-import com.mongodb.MongoClientSettings;
 import com.mongodb.MongoException;
 import com.mongodb.MongoTimeoutException;
 import com.mongodb.ReadConcern;
 import com.mongodb.ReadPreference;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
-import com.mongodb.connection.ClusterSettings;
 import org.bson.RawBsonDocument;
 
 import java.io.Closeable;
 import java.io.IOException;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 
-import static com.mongodb.internal.capi.MongoCryptOptionsHelper.createMongocryptdSpawnArgs;
+import static com.mongodb.internal.capi.MongoCryptHelper.createMongocryptdSpawnArgs;
+import static com.mongodb.internal.capi.MongoCryptHelper.createMongocryptdClientSettings;
 
 @SuppressWarnings("UseOfProcessBuilder")
 class CommandMarker implements Closeable {
@@ -52,24 +48,7 @@ class CommandMarker implements Closeable {
             } else {
                 processBuilder = null;
             }
-
-            String connectionString;
-
-            if (options.containsKey("mongocryptdURI")) {
-                connectionString = (String) options.get("mongocryptdURI");
-            } else {
-                connectionString = "mongodb://localhost:27020";
-            }
-
-            client = MongoClients.create(MongoClientSettings.builder()
-                    .applyConnectionString(new ConnectionString(connectionString))
-                    .applyToClusterSettings(new Block<ClusterSettings.Builder>() {
-                        @Override
-                        public void apply(final ClusterSettings.Builder builder) {
-                            builder.serverSelectionTimeout(1, TimeUnit.SECONDS);
-                        }
-                    })
-                    .build());
+            client = MongoClients.create(createMongocryptdClientSettings((String) options.get("mongocryptdURI")));
         }
     }
